Ajustes no código

+++ src/main/java/com/wagner/kroiss/domain/service/CadastroCozinhaService.java
@@ -5,6 +5,8 @@ import com.wagner.kroiss.domain.exceptions.EntidadeEmUsoException;
 import com.wagner.kroiss.domain.exceptions.EntidadeNaoEncontrada;
 import com.wagner.kroiss.domain.model.Cozinha;
 import com.wagner.kroiss.domain.repository.CozinhaRepository;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
@@ -14,7 +16,10 @@ import org.springframework.stereotype.Service;
 public class CadastroCozinhaService {
 
     @Autowired
-    private CozinhaRepository cozinhaRepository;
+    public CozinhaRepository cozinhaRepository;
+
+    private static final Logger logger = LoggerFactory.getLogger(CadastroCozinhaService.class);
+
 
     public Cozinha salvar(Cozinha cozinha) {
         return  cozinhaRepository.save(cozinha);
@@ -23,18 +28,19 @@ public class CadastroCozinhaService {
 
     public void excluir(Long cozinhaId) {
         try {
+            if (!cozinhaRepository.existsById(cozinhaId)) {
+                throw new EntidadeNaoEncontrada(
+                        String.format("Não existe uma cozinha com esse código %d", cozinhaId));
+            }
             cozinhaRepository.deleteById(cozinhaId);
-
         } catch (EmptyResultDataAccessException e) {
-            throw new EntidadeNaoEncontrada (
-                    String.format("Não existe uma cozinha com esse código %d", cozinhaId)
-            );
-        }
-
-        catch (DataIntegrityViolationException e) {
+            logger.error("Erro ao excluir cozinha com ID {}: {}", cozinhaId, e.getMessage());
+            throw new EntidadeNaoEncontrada(
+                    String.format("Não existe uma cozinha com esse código %d", cozinhaId));
+        } catch (DataIntegrityViolationException e) {
+            logger.error("Erro de integridade ao excluir cozinha com ID {}: {}", cozinhaId, e.getMessage());
             throw new EntidadeEmUsoException(
                     String.format("Cozinha de código %d não pode ser removida, pois está em uso", cozinhaId));
         }
-
     }
 }
